Split measurement into start and getting method
The measurement can be started the wait time spend doing something more
productive in order to speed up the code.
To this end, the measurement method is split into a conversion start
method which initializes the conversion and returns. The current time is
saved and the method pulling the data performs a time check against the
start time and the period needed for waiting.